Add files via upload
In today’s scenario, where online shopping has become a way of life, reviews hold a lot of value. A buyer tends to look at the ratings and the reviews given to the product before making a purchase. Looking at thousands of reviews can be a time-consuming task and choosing the right reviews for the prefect buy isn’t that easy. So, we aim to help customers choose the most significant reviews before making an online purchase.